third_party/blink/web_tests/html5lib/generated/run-webkit01-write-expected.txt

CONSOLE MESSAGE: PASS
CONSOLE MESSAGE: FOO<span>BAR</span>BAZ
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
../resources/webkit01.dat:
32

Test 32 of 52 in ../resources/webkit01.dat failed. Input:
<select><option>A<select><option>B<select><option>C<select><option>D<select><option>E<select><option>F<select><option>G<select>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "A"
|     <select>
|       <option>
|         "B"
|     <select>
|       <option>
|         "C"
|     <select>
|       <option>
|         "D"
|     <select>
|       <option>
|         "E"
|     <select>
|       <option>
|         "F"
|     <select>
|       <option>
|         "G"
|     <select>
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "A"
|     <option>
|       "B"
|       <select>
|         <option>
|           "C"
|     <option>
|       "D"
|       <select>
|         <option>
|           "E"
|     <option>
|       "F"
|       <select>
|         <option>
|           "G"
